+++ Alpha/AlphaTests/modelTest.swift
@@ -1,0 +1,37 @@
+//
+//  modelTest.swift
+//  AlphaTests
+//
+//  Created by hao yin on 2021/8/31.
+//
+
+import XCTest
+@testable import Alpha
+class modelTest: XCTestCase {
+
+    override func setUpWithError() throws {
+        // Put setup code here. This method is called before the invocation of each test method in the class.
+    }
+
+    override func tearDownWithError() throws {
+        // Put teardown code here. This method is called after the invocation of each test method in the class.
+    }
+
+    func testExample() throws {
+        print(MemoryLayout<test>.size,MemoryLayout<test>.stride,MemoryLayout<test>.alignment)
+        // This is an example of a functional test case.
+        // Use XCTAssert and related functions to verify your tests produce the correct results.
+    }
+
+    func testPerformanceExample() throws {
+        // This is an example of a performance test case.
+        self.measure {
+            // Put the code you want to measure the time of here.
+        }
+    }
+
+}
+public struct test{
+    @Col
+    var a:String = ""
+}
